refactor stringref API
